requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
channels==4.0.0
channels-redis==4.1.0
channels[daphne]
cryptography==41.0.4
daphne==4.0.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
service-identity==23.1.0
sqlparse==0.4.4
txaio==23.1.1
typing_extensions==4.8.0
websocket-client==1.6.3
websockets==11.0.3
